refactor(codexHooks): reorganize projectApplyPatchHunks and update documentation
- Moved the `projectApplyPatchHunks` function to a new directory structure under `hunkProjection` for better organization.
- Updated import paths in related files to reflect the new location of `projectApplyPatchHunks`.
- Enhanced the DETAIL.md documentation to clarify the boundary exemptions related to the new structure.
- Added comprehensive implementation of the `projectApplyPatchHunks` function to improve content projection logic without filesystem I/O.
